+++ src/CoreEx/Validation/IValidator.cs
@@ -1,0 +1,34 @@
+﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
+
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace CoreEx.Validation
+{
+    /// <summary>
+    /// Enables value validation.
+    /// </summary>
+    public interface IValidator
+    {
+        /// <summary>
+        /// Gets the <see cref="Type"/> for the value that is being validated.
+        /// </summary>
+        Type ValueType { get; }
+
+        /// <summary>
+        /// Validate the <paramref name="value"/>.
+        /// </summary>
+        /// <param name="value">The value to validate.</param>
+        /// <returns>The <see cref="IValidationResult"/>.</returns>
+        IValidationResult Validate(object value);
+
+        /// <summary>
+        /// Validate the <paramref name="value"/> asynchronously.
+        /// </summary>
+        /// <param name="value">The value to validate.</param>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>The <see cref="IValidationResult"/>.</returns>
+        Task<IValidationResult> ValidateAsync(object value, CancellationToken cancellationToken = default);
+    }
+}